fix submenu link classes, menu structure
